Version 1.1.1 Refactoring and fix of a bug.
Renamed Repository and repository.sqlite classes . Created two new methods in the RepositoryInMemory to reduce duplicated code. Fixed a bug in Controller were there was an extra printing of data that wasnt requested from the user. The bug was fixed by refactoring farther the Controller's switch inside the handler.  Created the class Actions to implement in the future part of Controller's switch.