Version 1.3.1 Connected the queryLoadDb and called in the methods to get the saved Events
This way we use one query to load the db data and then work on them based on the existed code we made in from reporisoryInMemory.
Also I commented the 2 tempList methods from the IRepository interface since the RepositoryDb needs to send the loaded list as well, there is a conflict to override them as different.