Version 1.4.0 Connecting it with the RepositoryDb , Changes on the queries , Bug to fix!
The RepositoryInMemory has been commented and in it's place the RepositoryDb has connected with the rest. Changed the queries for the ActionsToRun in order to take the data from the the DB without farther work on them.
BUG TO FIX : in the addMessage in the RepositoryDb , it doesnt allow a message to be added if its not a number ( though the variables of massage are Strings).