Version 1.4.4 Using both implementations. Fixes for repositoryDb.
The Controller now gives an object of IRepository which gets passed to the ActionsToRun method (runAction) in order to be able to use both implementations using the interface object. (There needs to be created a way for the user to choose which implementation to use. Also i have to find a way to get the specific class of the interface object inside the ActionsToRun in order to appropriately open-close the DB if needed.).
Fixed the overloading of past queries data in the RepositoryDb (removed the Arraylist<Event> before the initialising of the list repositoryDbList in the methods.)