Version 1.5.0 Bug fixed, Better architecture in both Repositories (changed in IRepository), Created JUnit5 Tests for each class, Finished Test for RepositoryInMemory.
Fixed a problem where the added message that was saved was only one word in size, the problem was in the ScannerImport class in method readNewMessage.
Refactored the method addNewMessage in all Repositories (inclusind the interface IRepository) classes in order to get an argument of Event instead of a event.getMessage and event.getTime as arguments. The actionsToRun method also sends an Event object now for the addNewMessage.
Created Test classes.
Implemented the RepositoryInMemory class (except the addNewMessage and didn't connect the @BeforeEach with the rest).