Version 2.0.1 JUnit for RepositoryDbTest, refactorring in Action, Event classes. Cleaning the code. More Tests implemented. There are bugs in the code to be fixed!
JUnit for RepositoryDbTest was created as well as more Tests for the rest of the classes where was possible. But there are some bugs in the code waiting to be fixed so some are commented for now.
Cleaned the code from older comments-code.
Removed the newEvent and newAction -as well as their sets- from their classes because they were not needed anywhere (the object Event and Action do the same thing).