Version 2.2.1 RepositoryDbTest : Fixed some Tests assertEquals. Changed the place where i delete the test-values that were added.
Changed the expected arrayListCompare  for Latest and Oldest methods in RepositoryDbTest.
Fixed some minor things in 2 more methods.
The deleting of the Test-values inserted is now called before the assertEquals so that the data is being restorred to their normal state even if the Test fails(with the assertEquals). (Otherwise when it would fail it would stop executing so the delete wouldnt run).