Version 2.2.3 Modifications for RepositoryDb and RepositoryInMemory addMessage classes.
The addMessage classes became boolean methods instead of void in order to return true or false in case something went wrong. Also the ActionsToRun now knows if it was a success or not and uses it aprropriatelly. By doing this the user will see a different message telling them if there was success or failure with their request.